Handle points added (on view). Coloring. Blah blah blah.

+++ fontcommons/src/main/java/spbstu/cg/fontcommons/PointUtils.java
@@ -9,11 +9,11 @@ import spbstu.cg.fontcommons.point.Point;
  */
 public class PointUtils {
 
-    public static double getSquaredDist(Point x, Point y) {
+    public static float getSquaredDist(Point x, Point y) {
         return (x.getX() - y.getX()) * (x.getX() - y.getX()) + (x.getY() - y.getY()) * (x.getY() - y.getY());
     }
 
-    public static double getSquaredDist(float x1, float y1, float x2, float y2) {
+    public static float getSquaredDist(float x1, float y1, float x2, float y2) {
         return (x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2);
     }
 }
